ページング処理追加

+++ src/main/java/com/example/demo/domain/model/Posts.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Posts {
 	
 	// 投稿番号(連番)
-	private int post_id;
+	private long post_id;
 	
 	// ユーザーID
 	private String user_id;
